feat: reimplement absolute mouse input on SDL

+++ 4J.Input/4J_Input.h
@@ -134,8 +134,8 @@ public:
 
 	//bool InputDetected(DWORD dwUserIndex,WCHAR *pwchInput);
 
-	float GetMouseX();
-	float GetMouseY();
+	int GetMouseX();
+	int GetMouseY();
 };
 
 // Singleton
